tests for myRemove function created.
