+++ m1/bloco_10/testes-automatizados-com-jest_1/.gitignore
@@ -1,0 +1,1 @@
+node_modules
