--- m1/bloco_10/testes-automatizados-com-jest_1/myRemove.js
@@ -1,17 +1,0 @@
-const assert = require('assert');
-
-function myRemove(arr, item) {
-  let newArr = [];
-  for (let index = 0; index < arr.length; index += 1) {
-    if (item !== arr[index]) {
-      newArr.push(arr[index]);
-    }
-  }
-  return newArr;
-}
-
-// implemente seus testes aqui
-// Verifique se a chamada myRemove([1, 2, 3, 4], 3) retorna o array esperado
-// Verifique se a chamada myRemove([1, 2, 3, 4], 3) não retorna o array [1, 2, 3, 4]
-// Verifique se o array passado por parâmetro não sofreu alterações
-// Verifique se a chamada myRemove([1, 2, 3, 4], 5) retorna o array esperado
